refactor: HttpRequestHandler 생성
RequestHandler.run 에서 발생하는 예외 처리 시 DataOutputStream 객체를 새로 생성하는 문제를 해결하기 위해 DataOutputStream 객체를 HttpRequestHandler에 주입시켜서 재사용하도록 변경
